Add Email auth page interstitial login page
This prevents the URL from being invalidated before the user is able to use the link.  Some software such as slack or firewall systems will access the link thus invalidating it before the user has been able to authenticate.  Now an intermediate page is displayed and the user has to click the login button on that page to authenticate.