Allow overriding MI email
The sync logic takes the user's primary email username and adds it to MI google domain to create the MI user.  However, this poses problems when we have multiple users with the same name. This PR adds the ability to manually override the MI google domain by adding a secondary email for the user.